Readme.md updated

+++ src/Screens/Level.js
@@ -18,7 +18,8 @@ class Level extends React.Component {
             data: [], gamingArray: {}, typeSelect: "", levelColor: '', levelName: '',
             buttonName: "Submit",
             sortOrder: '',
-            defineModule: false
+            defineModule: false,
+            sortOrderArray: []
         }
     }
 
@@ -42,34 +43,41 @@ class Level extends React.Component {
         let that = this;
         if (Object.keys(json).length > 0 && json['levelsMap'] != null && json['levelsMap'] != undefined) {
             let levelsMap = json['levelsMap'];
-            that.setState({ gamingArray: levelsMap })
+            let sortOrderArray = []
+            Object.keys(levelsMap).map((sortOrderObjects) => {
+                let sortOrderAllData = levelsMap[sortOrderObjects]
+                let sortOrderUniqData = sortOrderAllData.sortOrder
+                sortOrderArray.push(sortOrderUniqData.toString())
+            })
+            that.setState({ gamingArray: levelsMap, sortOrderArray })
         }
     }
 
     async deleteLevels(levelId) {
         let postJson = { sessionId: '1223', levelId: levelId };
         let responseData = await doConnect("deleteGameLevels", "POST", postJson);
-                if (responseData.response == 'Success') {
-                    toast.success(' Data deleted !', {
-                        position: "top-center",
-                        autoClose: 5000,
-                        hideProgressBar: false,
-                        closeOnClick: true,
-                        pauseOnHover: true,
-                        draggable: true,
-                        progress: undefined,
-                    });
+        if (responseData.response == 'Success') {
+            toast.success(' Module is deleted successfully!', {
+                position: "top-center",
+                autoClose: 5000,
+                hideProgressBar: false,
+                closeOnClick: true,
+                pauseOnHover: true,
+                draggable: true,
+                progress: undefined,
+            });
 
-                    this.getLevels();
+            this.getLevels();
 
-                } else {
-                    alert(responseData.response);
-                }
+        } else {
+            alert(responseData.response);
+        }
     }
 
 
     async createLevel() {
-        const { levelName, levelColor, selectedOption, sortOrder } = this.state;
+        const { levelName, levelColor, selectedOption, sortOrder, sortOrderArray
+        } = this.state;
         var validLength = levelColor ? levelColor.search("#") : 0;
         if (levelName.length == 0) {
             this.setState({ levelNameValidate: 'Please Enter Value' })
@@ -99,7 +107,11 @@ class Level extends React.Component {
             this.setState({ levelImageValidate: '' })
         }
         if (!Number(sortOrder)) {
-            this.setState({ sortOrderValidate: 'Please Sort Order' })
+            this.setState({ sortOrderValidate: 'Please enter sort order' })
+            return false
+        }
+        else if (sortOrderArray.includes(sortOrder.toString())) {
+            this.setState({ sortOrderValidate: 'This order number is already available' })
             return false
         }
         else {
@@ -112,7 +124,7 @@ class Level extends React.Component {
             var json = responseData;
             var response = json.response;
             if (response == 'Success') {
-                toast.success('Added data !', {
+                toast.success('Module is created Successfully !', {
                     position: "top-center",
                     autoClose: 5000,
                     hideProgressBar: false,
@@ -121,7 +133,7 @@ class Level extends React.Component {
                     draggable: true,
                     progress: undefined,
                 });
-                this.setState({ buttonName: 'Submit', selectedOption: {}, typeSelect: '', idvalue: "", levelColor: "", levelName: "" })
+                this.setState({ buttonName: 'Submit', selectedOption: {}, typeSelect: '', idvalue: "", levelColor: "", levelName: "", defineModule: false })
                 this.getLevels();
             } else {
                 alert(response);
@@ -130,7 +142,7 @@ class Level extends React.Component {
     }
 
     async updateLevels() {
-        const { levelName, levelColor, sortOrder } = this.state;
+        const { levelName, levelColor, sortOrder, sortOrderArray } = this.state;
         let validLength = levelColor ? levelColor.search("#") : 0;
         if (levelName.length == 0) {
             this.setState({ levelNameValidate: 'Please Enter Value' })
@@ -148,6 +160,17 @@ class Level extends React.Component {
         } else {
             this.setState({ levelColorValidate: '' })
         }
+        if (!Number(sortOrder)) {
+            this.setState({ sortOrderValidate: 'Please enter sort order' })
+            return false
+        }
+        else if (sortOrderArray.includes(sortOrder.toString())) {
+            this.setState({ sortOrderValidate: 'This order number is already available' })
+            return false
+        }
+        else {
+            this.setState({ sortOrderValidate: '' })
+        }
         let found = this.state.levelArray.findIndex((a) =>
             a.id === this.state.idvalue
         )
@@ -160,7 +183,7 @@ class Level extends React.Component {
             let postJson = { levelId: this.state.idvalue, name: levelName, color: levelColor, sessionId: '1223', image: this.state.selectedOption.json, sortOrder: Number(sortOrder) };
             let responseData = await doConnect("updateGameLevel", "POST", postJson);
             if (responseData.response == 'Success') {
-                toast.success('Updated data !', {
+                toast.success('Module is updated Successfully !', {
                     position: "top-center",
                     autoClose: 5000,
                     hideProgressBar: false,
@@ -169,7 +192,7 @@ class Level extends React.Component {
                     draggable: true,
                     progress: undefined,
                 });
-                this.setState({ buttonName: 'Submit', selectedOption: {}, typeSelect: '', idvalue: "", levelColor: "", levelName: "" })
+                this.setState({ buttonName: 'Submit', selectedOption: {}, typeSelect: '', idvalue: "", levelColor: "", levelName: "", defineModule: false })
 
                 this.getLevels();
             } else {
@@ -182,7 +205,7 @@ class Level extends React.Component {
 
     render() {
 
-        const { buttonName, defineModule } = this.state;
+        const { buttonName, defineModule, sortOrder } = this.state;
 
         const columns = [
             {
@@ -217,7 +240,7 @@ class Level extends React.Component {
                     let image = row.image;
                     return <div style={{ padding: 10 }}>
 
-                        <button id={row.id} className="btn btn-info" onClick={() => {
+                        <button id={row.id} className="btn btnc" onClick={() => {
                             this.props.history.push('/' + MyConstant.keyList.projectUrl + '/module-manager-ide/' + row.id)
                         }}>Manage</button>
 
@@ -255,7 +278,8 @@ class Level extends React.Component {
 
                             let levelName = this.state.gamingArray[e.target.id].name
                             let levelColor = this.state.gamingArray[e.target.id].color
-                            this.setState({ defineModule: true,buttonName: 'Update', selectedOption: object, typeSelect: 'Edit', idvalue: e.target.id, levelColor, levelName })
+                            let sortOrder = this.state.gamingArray[e.target.id].sortOrder
+                            this.setState({ defineModule: true, buttonName: 'Update', selectedOption: object, typeSelect: 'Edit', idvalue: e.target.id, levelColor, levelName, sortOrder })
                         }}>Edit</button>
                     </div>,
             },
@@ -268,7 +292,9 @@ class Level extends React.Component {
                     >
                         <div style={{ fontWeight: 700 }}></div>
                         <button id={row.id} className="btn btn-danger" onClick={(e) => {
-                            this.deleteLevels(e.target.id)
+                            if (window.confirm("Please confirm to Delete this Module")) {
+                                this.deleteLevels(e.target.id)
+                            }
                         }}>Delete</button>  </div>,
             }
         ];
@@ -301,7 +327,11 @@ class Level extends React.Component {
                                             <div className="float-right">
                                                 <button className="btn btn-success" onClick={() => {
                                                     this.setState({
-                                                        defineModule: true
+                                                        defineModule: true,
+                                                        sortOrder: "",
+                                                        levelColor: "",
+                                                        selectedOption: {},
+                                                        levelName: ""
                                                     })
                                                 }}>Add Module</button>
                                             </div>
@@ -370,31 +400,39 @@ class Level extends React.Component {
                                                                                 }}>{buttonName}</button>
                                                                                 <button type="button" className="btn btn-danger" onClick={() => {
                                                                                     this.setState({
-                                                                                        defineModule: false
+                                                                                        defineModule: false,
+                                                                                        levelNameValidate: "",
+                                                                                        levelColorValidate: "",
+                                                                                        levelImageValidate: "",
+                                                                                        sortOrderValidate: ""
                                                                                     })
                                                                                 }}>Cancel</button>
                                                                             </div>
                                                                         </div>
-                                                                        </React.Fragment> : <React.Fragment>
+                                                                    </React.Fragment> : <React.Fragment>
                                                                         <div className="row">
                                                                             <div className="col-12 text-right">
                                                                                 <button type="button" className={'btn btn-success '} onClick={() => {
                                                                                     this.updateLevels()
                                                                                 }}>{buttonName}
                                                                                 </button>
-                                                                            <button type="button" className="btn btn-danger" onClick={() => {
-                                                                                this.setState({
-                                                                                    defineModule: false,
-                                                                                    selectedOption: {}, 
-                                                                                    typeSelect: '', 
-                                                                                    idvalue: "", 
-                                                                                    levelColor: "", 
-                                                                                    levelName: ""
-                                                                                })
-                                                                            }}>Cancel</button>
+                                                                                <button type="button" className="btn btn-danger" onClick={() => {
+                                                                                    this.setState({
+                                                                                        defineModule: false,
+                                                                                        selectedOption: {},
+                                                                                        typeSelect: '',
+                                                                                        idvalue: "",
+                                                                                        levelColor: "",
+                                                                                        levelName: "",
+                                                                                        levelNameValidate: "",
+                                                                                        levelColorValidate: "",
+                                                                                        levelImageValidate: "",
+                                                                                        sortOrderValidate: ""
+                                                                                    })
+                                                                                }}>Cancel</button>
+                                                                            </div>
                                                                         </div>
-                                                                    </div>
-                                                                </React.Fragment>
+                                                                    </React.Fragment>
                                                                 }
                                                             </div>
                                                         </div>
